requirements.txt
annotated-types==0.7.0
anyio==4.9.0
black==25.1.0
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
coverage==7.8.0
# Editable Git install with no remote (cybercare==0.1.0)
-e /home/vastas/gitrepos/cybercare
distlib==0.3.9
exceptiongroup==1.2.2
fastapi==0.115.12
filelock==3.18.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
identify==2.6.9
idna==3.10
iniconfig==2.1.0
isort==6.0.1
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.5.0
pre_commit==4.2.0
psycopg2-binary==2.9.10
pydantic==2.11.3
pydantic_core==2.33.1
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-cov==6.1.1
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
starlette==0.46.2
tomli==2.2.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.1
virtualenv==20.30.0
